Moved prior's computing routines from definition to inference code base
