+++ probeye/inference/taralli_/__init__.py
@@ -1,2 +1,3 @@
 # module imports
 from probeye.inference.taralli_ import solver
+from probeye.inference.taralli_ import priors
